october 2022

+++ install.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
+if (( $EUID != 0 )); then
+    echo "Please run as root"
+    exit
+fi
+
 cd ~/.dotfiles
 
 # Install packages
-sudo pacman -Syu --noconfirm paru
+pacman -Syu --noconfirm paru
 paru -Rns --noconfirm firefox
 cat packages.lst | xargs paru -Syu --needed --noconfirm
 
 # to copy dolphin layout files
-mkdir ~/.local/share/kxmlgui5 2>/dev/null
+sudo -u $SUDO_USER mkdir ~/.local/share/kxmlgui5 2>/dev/null
 
 # remove useless dirs
 rm -r ~/Templates 2>/dev/null
@@ -20,7 +25,7 @@ for file in .*; do
     if [[ "$file" == "." ]] || [[ "$file" == ".." ]]; then
         continue
     fi
-    ln -sf ~/.dotfiles/home/$file ~/$file
+    sudo -u $SUDO_USER ln -sf ~/.dotfiles/home/$file ~/$file
 done
 cd ..
 
@@ -31,80 +36,80 @@ for dotDir in */; do
     for dir in $dotDir*; do
         unescapedDir=$(echo $dir | sed "s/dot/./g" | sed "s/_/\//g")
         rm -rf ~/$unescapedDir 2>/dev/null
-        ln -sf ~/.dotfiles/$dir ~/$unescapedDir 2>/dev/null \
+        sudo -u $SUDO_USER ln -sf ~/.dotfiles/$dir ~/$unescapedDir 2>/dev/null \
           || echo "$dir cannot be transfered" # message when error
     done
 done
 
 for file in $(ls myscripts); do
-    sudo ln -sf ~/.dotfiles/myscripts/$file /usr/local/bin/$file
+    ln -sf ~/.dotfiles/myscripts/$file /usr/local/bin/$file
 done
 
 # Copy specifc settings
 
 # libinput gestures
-sudo gpasswd -a $USER input
-libinput-gestures-setup autostart
+gpasswd -a $USER input
+sudo -u $SUDO_USER libinput-gestures-setup autostart
 
 # docker
-sudo gpasswd -a $user docker
-sudo systemctl enable docker.service
+gpasswd -a $user docker
+systemctl enable docker.service
 
 # vim
-mkdir -p ~/.vim/undodir 2>/dev/null
-vim -c "PlugInstall" -c "PlugClean" -c "qa!"
-ln -sf ~/.dotfiles/vim/coc-settings.json ~/.vim/coc-settings.json
-ln -sf ~/.dotfiles/vim/fzf-gruvbox.config ~/.vim/fzf-gruvbox.config
-sudo npm i -g bash-language-server
+sudo -u $SUDO_USER mkdir -p ~/.vim/undodir 2>/dev/null
+sudo -u $SUDO_USER vim -c "PlugInstall" -c "PlugClean" -c "qa!"
+sudo -u $SUDO_USER ln -sf ~/.dotfiles/vim/coc-settings.json ~/.vim/coc-settings.json
+sudo -u $SUDO_USER ln -sf ~/.dotfiles/vim/fzf-gruvbox.config ~/.vim/fzf-gruvbox.config
+npm i -g bash-language-server
 
 # vscode settings
-mkdir -p ~/.config/Code/User 2>/dev/null
-ln -sf ~/.dotfiles/vscode/settings.json ~/.config/Code/User/settings.json
+sudo -u $SUDO_USER mkdir -p ~/.config/Code/User 2>/dev/null
+sudo -u $SUDO_USER ln -sf ~/.dotfiles/vscode/settings.json ~/.config/Code/User/settings.json
 
 # jetbrains settings
 programs=("clion" "intellij" "pycharm")
 for nickname in ${programs[@]}; do
     timeout 1s /bin/$nickname* # open the app to create the config folder
     program=$(ls ~/.config/JetBrains | grep -i $nickname)
-    mkdir -p ~/.config/JetBrains/$program/options 2>/dev/null
+    sudo -u $SUDO_USER mkdir -p ~/.config/JetBrains/$program/options 2>/dev/null
     for file in $(ls ~/.dotfiles/jetbrains/); do
-        ln -sf ~/.dotfiles/jetbrains/$file ~/.config/JetBrains/$program/options/$file
+        sudo -u $SUDO_USER ln -sf ~/.dotfiles/jetbrains/$file ~/.config/JetBrains/$program/options/$file
     done
 done
 
 # add zsh plugins
-sudo git clone https://github.com/zsh-users/zsh-syntax-highlighting.git /usr/share/zsh/plugins/zsh-syntax-highlighting/
-sudo git clone https://github.com/zsh-users/zsh-history-substring-search.git /usr/share/zsh/plugins/zsh-history-substring-search/
-sudo git clone https://github.com/zsh-users/zsh-autosuggestions.git /usr/share/zsh/plugins/zsh-autosuggestions/
+git clone https://github.com/zsh-users/zsh-syntax-highlighting.git /usr/share/zsh/plugins/zsh-syntax-highlighting/
+git clone https://github.com/zsh-users/zsh-history-substring-search.git /usr/share/zsh/plugins/zsh-history-substring-search/
+git clone https://github.com/zsh-users/zsh-autosuggestions.git /usr/share/zsh/plugins/zsh-autosuggestions/
 
 # transfer wallpapers
-ln -sf ~/.dotfiles/wallpapers ~/Pictures/wallpapers
-ln -sf ~/.dotfiles/media/abstergo-transparent-small.png ~/Pictures/abstergo-transparent-small.png
-plasma-apply-wallpaperimage ~/Pictures/wallpapers/wallhaven-136m9w.png
+sudo -u $SUDO_USER ln -sf ~/.dotfiles/wallpapers ~/Pictures/wallpapers
+sudo -u $SUDO_USER ln -sf ~/.dotfiles/media/abstergo-transparent-small.png ~/Pictures/abstergo-transparent-small.png
+sudo -u $SUDO_USER plasma-apply-wallpaperimage ~/Pictures/wallpapers/wallhaven-136m9w.png
 
 # transfer sddm background
 # (doesn't work with ln)
-sudo cp ~/.dotfiles/sddm/theme.conf /usr/share/sddm/themes/breeze/
-sudo cp ~/.dotfiles/sddm/theme.conf.user /usr/share/sddm/themes/breeze/
-sudo cp ~/.dotfiles/wallpapers/sunset-in-the-mountains-illustration_3840x2160_xtrafondos.com.jpg /usr/share/sddm/themes/breeze/
+cp ~/.dotfiles/sddm/theme.conf /usr/share/sddm/themes/breeze/
+cp ~/.dotfiles/sddm/theme.conf.user /usr/share/sddm/themes/breeze/
+cp ~/.dotfiles/wallpapers/sunset-in-the-mountains-illustration_3840x2160_xtrafondos.com.jpg /usr/share/sddm/themes/breeze/
 
 # Apply icon theme
-/usr/lib/plasma-changeicons ~/.local/share/icons/kora
+sudo -u $SUDO_USER /usr/lib/plasma-changeicons ~/.local/share/icons/kora
 
 # Make fonts available on system (p.ex filename in Dolphin)
-sudo ln -sf ~/.dotfiles/local.conf  /etc/fonts/local.conf
+ln -sf ~/.dotfiles/local.conf  /etc/fonts/local.conf
 
 # save firefox settings
 saveFirefoxData () {
     # if too big, clear cache before
     oldPath=$(pwd)
     cd ~/
-    tar -jcvf dotmozilla.tar.bz2 .mozilla
-    mv dotmozilla.tar.bz2 ~/.dotfiles/
+    sudo -u $SUDO_USER tar -jcvf dotmozilla.tar.bz2 .mozilla
+    sudo -u $SUDO_USER mv dotmozilla.tar.bz2 ~/.dotfiles/
     cd ~/.dotfiles
     read -sp "You will have to enter the key to encrypt Firefox profile [Enter]"
     echo
-    gpg -c dotmozilla.tar.bz2
+    sudo -u $SUDO_USER gpg -c dotmozilla.tar.bz2
     rm dotmozilla.tar.bz2
     cd $oldPath
 }
@@ -116,41 +121,40 @@ restoreFirefoxData () {
     rm -rf ~/.mozilla 2>/dev/null
     read -sp "You will have to enter the key to decrypt Firefox profile [Enter]"
     echo
-    gpg ~/.dotfiles/dotmozilla.tar.bz2.gpg
-    mv ~/.dotfiles/dotmozilla.tar.bz2 ./
-    tar -xvf ~/dotmozilla.tar.bz2
+    sudo -u $SUDO_USER gpg ~/.dotfiles/dotmozilla.tar.bz2.gpg
+    sudo -u $SUDO_USER mv ~/.dotfiles/dotmozilla.tar.bz2 ./
+    sudo -u $SUDO_USER tar -xvf ~/dotmozilla.tar.bz2
     rm ~/dotmozilla.tar.bz2
     cd $oldPath
 }
-restoreFirefoxData
 
 # Smooth scroll firefox
-echo export MOZ_USE_XINPUT2=1 | sudo tee /etc/profile.d/use-xinput2.sh
+sudo -u $SUDO_USER echo export MOZ_USE_XINPUT2=1 | sudo tee /etc/profile.d/use-xinput2.sh
 
-# change shell for zsh
-chsh -s /bin/zsh
 
 # Set too many errors password time to 10s
-sudo sed -i '45 a\unlock_time = 10' /etc/security/faillock.conf
+sed -i '45 a\unlock_time = 10' /etc/security/faillock.conf
 
 # optimus manager settings
 systemctl enable optimus-manager.service
 
-# timeshift settings
-sudo ln -sf ~/.dotfiles/timeshift.json /etc/timeshift/timeshift.json
-
 # set grub chose time to 0 seconds
-sudo vim /etc/default/grub -u NONE -c "/GRUB_TIMEOUT" -c "s/[0-9]\+/0" -c "wq"
-sudo grub-mkconfig -o /boot/grub/grub.cfg
+vim /etc/default/grub -u NONE -c "/GRUB_TIMEOUT" -c "s/[0-9]\+/0" -c "wq"
+grub-mkconfig -o /boot/grub/grub.cfg
 
 # Configure wireshark
-sudo chmod +x /usr/bin/dumpcap
+chmod +x /usr/bin/dumpcap
 
-sudo systemctl enable bluetooth
+systemctl enable bluetooth
+
+# change shell for zsh
+sudo -u $SUDO_USER chsh -s /bin/zsh
+
+restoreFirefoxData
 
 echo Everything is done !
 read -p "Do you want to reboot to apply config ?[y/n]: " userEntry
 if [[ "$userEntry" == "y" ]]; then
-    sudo shutdown -r now
+    shutdown -r now
 fi
